herle_inventarios/requirements.txt
Django==1.9.5
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.4.4
ecdsa==0.13
feedparser==5.1.3
httplib2==0.8
numpy==1.8.2
oauthlib==0.6.1
psycopg2==2.6.2
pycrypto==2.6.1
pyparsing==2.0.1
pytz==2016.4
pyxdg==0.25
requests==2.2.1
selenium==2.53.2
six==1.5.2
urllib3==1.7.1
wheel==0.26.0
